Add files via upload
Problem is :- "Convert the OS based program into a menu driven program using Python Code which will execute the required user query when user will give the input as a text."

This program is menu driven program in which we give any task to perform and it will perform that task for us, we give the command as text and it perform very well and very fast. the condition is that program had to be installed in the pc then it will perform well.
And the other thing is i focused on its  user interface so that it is user friendly, anyone can easily perform the task that is available in the menu of this program.